requirements.txt
openai-whisper
customtkinter